Refactoring progress
