+++ cpp/src/parquet/column_writer-test.cc
@@ -23,11 +23,11 @@
 #include "parquet/column_writer.h"
 #include "parquet/metadata.h"
 #include "parquet/properties.h"
+#include "parquet/statistics.h"
 #include "parquet/test-specialization.h"
 #include "parquet/test-util.h"
 #include "parquet/thrift.h"
 #include "parquet/types.h"
-#include "parquet/util/comparison.h"
 #include "parquet/util/memory.h"
 
 namespace parquet {
